backend/requirements.txt
aioprometheus==23.12.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cssselect2==0.8.0
cymem==2.0.11
DAWG-Python==0.7.2
DAWG2-Python==0.9.0
docopt==0.6.2
docutils==0.21.2
fastapi==0.115.6
fonttools==4.57.0
greenlet==3.2.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
jellyfish==1.2.0
Jinja2==3.1.6
jose==1.0.0
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
preshed==3.0.9
psycopg2-binary==2.9.10
pyahocorasick==2.1.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.6.1
pydantic_core==2.33.1
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.3
pymorphy3-dicts-ru==2.4.417150.4580142
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyxDamerauLevenshtein==1.8.0
quantile-python==1.1
RapidFuzz==3.13.0
reportlab==4.4.0
requests==2.32.3
rich==14.0.0
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl#sha256=69978d47b43e2c4f329bebdb155e8e9d3861bba1a58ba25551419dae7d7e07fc
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqladmin==0.20.1
SQLAlchemy==2.0.36
SQLAlchemy-JSONField==1.0.2
sqlalchemy-stubs==0.4
srsly==2.5.1
starlette==0.41.3
textdistance==4.6.3
thinc==8.3.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.15.2
types-ujson==5.10.0.20250326
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.32.1
uvloop==0.21.0
wasabi==1.1.3
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
wrapt==1.17.2
WTForms==3.1.2
zopfli==0.2.3.post1
black
isort
flake8